api/requirements.txt
Flask==2.0.1
python-docx==0.8.11
pdfkit==0.6.1
